uteoj/requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.4.0
billiard==3.5.0.5
bleach==3.3.0
celery==4.2.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-markdownify==0.9.0
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
kombu==4.6.11
Markdown==3.3.4
mysql==0.0.3
mysqlclient==2.0.3
packaging==21.0
Pillow==8.3.1
prompt-toolkit==3.0.19
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
zipfile38==0.0.3
